+++ Tests/OAuth2ImplicitGrant_tests.swift
@@ -38,7 +38,7 @@ class OAuth2ImplicitGrantTests: XCTestCase
 		XCTAssertEqual(oauth.authURL, URL(string: "https://auth.ful.io")!, "Must init `authorize_uri`")
 	}
 	
-	func testReturnURLHandling() {
+	func testDeprecatedReturnURLHandling() {
 		let oauth = OAuth2ImplicitGrant(settings: [
 			"client_id": "abc",
 			"authorize_uri": "https://auth.ful.io",
@@ -50,8 +50,8 @@ class OAuth2ImplicitGrantTests: XCTestCase
 			XCTAssertNotNil(error, "Error message expected")
 			XCTAssertEqual((error as! OAuth2Error), OAuth2Error.invalidRedirectURL(""))
 		}
-		oauth.afterAuthorizeOrFailure = { wasFailure, error in
-			XCTAssertTrue(wasFailure)
+		oauth.afterAuthorizeOrFail = { authParameters, error in
+			XCTAssertNil(authParameters)
 			XCTAssertNotNil(error, "Error message expected")
 		}
 		oauth.context._state = "ONSTUH"
@@ -118,8 +118,10 @@ class OAuth2ImplicitGrantTests: XCTestCase
 		oauth.onFailure = { error in
 			XCTAssertTrue(false, "Should not call this")
 		}
-		oauth.afterAuthorizeOrFailure = { wasFailure, error in
-			XCTAssertFalse(wasFailure)
+		oauth.afterAuthorizeOrFail = { authParameters, error in
+			XCTAssertNotNil(authParameters, "Expecting non-nil auth dict")
+			XCTAssertTrue(authParameters?.count > 2, "Expecting non-empty auth dict")
+			print("---->  \(authParameters)")
 			XCTAssertNil(error, "No error message expected")
 		}
 		oauth.handleRedirectURL(URL(string: "https://auth.ful.io#token_type=bearer&access_token=abc&state=\(oauth.context.state)&expires_in=3599")!)
@@ -131,6 +133,116 @@ class OAuth2ImplicitGrantTests: XCTestCase
 		// success 2
 		oauth.onFailure = { error in
 			XCTAssertTrue(false, "Should not call this")
+		}
+		oauth.handleRedirectURL(URL(string: "https://auth.ful.io#token_type=bearer&access_token=abc&state=\(oauth.context.state)")!)
+		XCTAssertNotNil(oauth.accessToken, "Must have an access token")
+		XCTAssertEqual(oauth.accessToken!, "abc")
+		XCTAssertNil(oauth.accessTokenExpiry)
+		XCTAssertTrue(oauth.hasUnexpiredAccessToken())
+	}
+	
+	func testReturnURLHandling() {
+		let oauth = OAuth2ImplicitGrant(settings: [
+			"client_id": "abc",
+			"authorize_uri": "https://auth.ful.io",
+			"keychain": false,
+		])
+		
+		// Empty redirect URL
+		oauth.didAuthorizeOrFail = { authParameters, error in
+			XCTAssertNil(authParameters, "Nil auth dict expected")
+			XCTAssertNotNil(error, "Error message expected")
+			XCTAssertEqual((error as! OAuth2Error), OAuth2Error.invalidRedirectURL(""))
+		}
+		oauth.afterAuthorizeOrFail = { authParameters, error in
+			XCTAssertNil(authParameters, "Nil auth dict expected")
+			XCTAssertNotNil(error, "Error message expected")
+		}
+		oauth.context._state = "ONSTUH"
+		oauth.handleRedirectURL(URL(string: "")!)
+		XCTAssertNil(oauth.accessToken, "Must not have an access token")
+		
+		// No params in redirect URL
+		oauth.didAuthorizeOrFail = { authParameters, error in
+			XCTAssertNil(authParameters, "Nil auth dict expected")
+			XCTAssertNotNil(error, "Error message expected")
+			XCTAssertEqual((error as! OAuth2Error), OAuth2Error.invalidRedirectURL("https://auth.ful.io"))
+		}
+		oauth.handleRedirectURL(URL(string: "https://auth.ful.io")!)
+		XCTAssertNil(oauth.accessToken, "Must not have an access token")
+		
+		// standard error
+		oauth.context._state = "ONSTUH"		// because it has been reset
+		oauth.didAuthorizeOrFail = { authParameters, error in
+			XCTAssertNil(authParameters, "Nil auth dict expected")
+			XCTAssertNotNil(error, "Error message expected")
+			XCTAssertEqual((error as! OAuth2Error), OAuth2Error.accessDenied)
+			XCTAssertEqual((error as! OAuth2Error).description, "The resource owner or authorization server denied the request.")
+		}
+		oauth.handleRedirectURL(URL(string: "https://auth.ful.io#error=access_denied")!)
+		XCTAssertNil(oauth.accessToken, "Must not have an access token")
+		
+		// explicit error
+		oauth.context._state = "ONSTUH"		// because it has been reset
+		oauth.didAuthorizeOrFail = { authParameters, error in
+			XCTAssertNil(authParameters, "Nil auth dict expected")
+			XCTAssertNotNil(error, "Error message expected")
+			XCTAssertNotEqual((error as! OAuth2Error), OAuth2Error.generic("Not good"))
+			XCTAssertEqual((error as! OAuth2Error), OAuth2Error.responseError("Not good"))
+			XCTAssertEqual((error as! OAuth2Error).description, "Not good")
+		}
+		oauth.handleRedirectURL(URL(string: "https://auth.ful.io#error_description=Not+good")!)
+		XCTAssertNil(oauth.accessToken, "Must not have an access token")
+		
+		// no token type
+		oauth.context._state = "ONSTUH"		// because it has been reset
+		oauth.didAuthorizeOrFail = { authParameters, error in
+			XCTAssertNil(authParameters, "Nil auth dict expected")
+			XCTAssertNotNil(error, "Error message expected")
+			XCTAssertEqual((error as! OAuth2Error), OAuth2Error.noTokenType)
+		}
+		oauth.handleRedirectURL(URL(string: "https://auth.ful.io#access_token=abc&state=\(oauth.context.state)")!)
+		XCTAssertNil(oauth.accessToken, "Must not have an access token")
+		
+		// unsupported token type
+		oauth.context._state = "ONSTUH"		// because it has been reset
+		oauth.didAuthorizeOrFail = { authParameters, error in
+			XCTAssertNil(authParameters, "Nil auth dict expected")
+			XCTAssertNotNil(error, "Error message expected")
+			XCTAssertEqual((error as! OAuth2Error), OAuth2Error.unsupportedTokenType("Only “bearer” token is supported, but received “helicopter”"))
+		}
+		oauth.handleRedirectURL(URL(string: "https://auth.ful.io#token_type=helicopter&access_token=abc&state=\(oauth.context.state)")!)
+		XCTAssertNil(oauth.accessToken, "Must not have an access token")
+		
+		// Invalid state
+		oauth.context._state = "ONSTUH"		// because it has been reset
+		oauth.didAuthorizeOrFail = { authParameters, error in
+			XCTAssertNil(authParameters, "Nil auth dict expected")
+			XCTAssertNotNil(error, "Error message expected")
+			XCTAssertEqual((error as! OAuth2Error), OAuth2Error.invalidState)
+		}
+		oauth.handleRedirectURL(URL(string: "https://auth.ful.io#token_type=bearer&access_token=abc&state=ONSTOH")!)
+		XCTAssertNil(oauth.accessToken, "Must not have an access token")
+		
+		// success 1
+		oauth.didAuthorizeOrFail = { authParameters, error in
+			XCTAssertNotNil(authParameters, "auth parameters expected")
+			XCTAssertNil(error, "No error message expected")
+		}
+		oauth.afterAuthorizeOrFail = { authParameters, error in
+			XCTAssertNotNil(authParameters, "auth parameters expected")
+			XCTAssertNil(error, "No error message expected")
+		}
+		oauth.handleRedirectURL(URL(string: "https://auth.ful.io#token_type=bearer&access_token=abc&state=\(oauth.context.state)&expires_in=3599")!)
+		XCTAssertNotNil(oauth.accessToken, "Must have an access token")
+		XCTAssertEqual(oauth.accessToken!, "abc")
+		XCTAssertNotNil(oauth.accessTokenExpiry)
+		XCTAssertTrue(oauth.hasUnexpiredAccessToken())
+		
+		// success 2
+		oauth.didAuthorizeOrFail = { authParameters, error in
+			XCTAssertNotNil(authParameters, "auth parameters expected")
+			XCTAssertNil(error, "No error message expected")
 		}
 		oauth.handleRedirectURL(URL(string: "https://auth.ful.io#token_type=bearer&access_token=abc&state=\(oauth.context.state)")!)
 		XCTAssertNotNil(oauth.accessToken, "Must have an access token")